pyproject.toml
[project]
name = "acp-mcpdoc-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "acp-sdk>=0.8.1",
    "beeai-framework[duckduckgo,wikipedia]>=0.1.19",
    "langchain-aws>=0.2.22",
    "langchain-mcp-adapters>=0.0.9",
    "langchain-ollama>=0.3.2",
    "langgraph>=0.4.1",
    "mcpdoc>=0.0.8",
    "pydantic-settings>=2.9.1",
]
